GlobalMechanism filter_dict improvements
A new keyword has been added to global mechanisms: recursive_species_filtering = True/False.

This allows for filter_dict keywords to be searched recursively through the species inside ComplexSpecies or not.

A new recursive method , get_species, was added to ComplexSpecies in order to do this.

This functionality is tested in a new test file, test_global_mechanism.py.

This functionality has also been documented in example notebook number 6.