Changes to make ExpressionMixtures work
The issue is that ExpressionMixtures were replacing Species. In retrospect, this is not the rigth way for them to work. Now they can use Mechanisms normally by just setting all transcripts in DNA assembly to be None (if Protein is not None). The mechanisms now switch between Expression and Transcription in this case.

At this point, all unit tests work.