add new parameter classes (#124)
* New Parameter Classes

Parameters are now stored in a ParameterDatabase populated by ParameterEntries. These new objects have required changes to how Mixtures and Components find, store, and load parameters (and initial conditions).

* ParameterKeys are namedtuples

ParameterKeys in ParameterDatabase are namedtuples.

* Associated Tests for Parameter objects have been made and tests for Mixtures and Components have been updated.

* For now, Parameters are returned as numbers (instead of ParameterEntries) - this will need to be undone when new parameter classes are added with new reaction classes.

* add checks to string parsing in Parameter

Parameter object accepts parameters values from strings

* Changes to make ExpressionMixtures work

The issue is that ExpressionMixtures were replacing Species. In retrospect, this is not the rigth way for them to work. Now they can use Mechanisms normally by just setting all transcripts in DNA assembly to be None (if Protein is not None). The mechanisms now switch between Expression and Transcription in this case.

* Combinatorial Testing and Fixes

Added tests to tests combinations of Components and Mixtures.

* Added .replace_species unit tests

To get coverage in chemical_reaction_network.py back to where it was previously, I added unit tests for ChemicalReactionNetwork, Species, ComplexSpecies OrderedComplexSpecies and Reaction .replace_species(old_s, new_s) function. This is no longer used in the code, but could be useful I imagine.


Co-authored-by: William Poole <wpoole@caltech.edu>
Co-authored-by: zoltuz <zoltuz@gmail.com>