add new DNA construction (#102)
This commit adds a separate path to set up DNA assemblies where, for example, multiple binding sites are present on the DNA or other complicated biocircuits operations is needed (i.e. integrase in the future). This path is recommended when biocircuit is already built in the lab and a companion in silico model is needed. For exploratory biocircuits, DNAassembly class is still recommended.   

* DNAComplexSpecies

* DNA binding site and dna_to_bind

* OrderedPolymer new way to do DNA_construct

* rerunning update_species happens in dna_construct

* Added tests for DNA assembly and GlobalMechanisms

Additionally propagated some of the changes into DNAassembly - importantly when overwriting Mechanisms, DNAassembly never overwrites the mechanisms of its promoter and rbs, these must be accessed individually.



Co-authored-by: William Poole <wpoole@caltech.edu>