refactor CRN simulator initial condition dictionary to accept Species (#208)
* refactor initial_concentration dictionary to support Species type

CRN simulators need an initial_concentration dictionary which has a pair of Species and numeric values.
So far user had to provide a string representation of a Species. Now Species type is also allowed.

* refactor initial_concentration dictionary to support Species type

CRN simulators need an initial_concentration dictionary which has a pair of Species and numeric values.
So far user had to provide a string representation of a Species. Now Species type is also allowed.