refactor Enzyme class
MultiEnzyme and Enzyme classes are merged because they had the same functionality. The signature of Enzyme has changed to subtrates (plural) and products (plural), but it accepts single string, Species or Component not just lists of those. Example notebooks are adjusted to the new signature.
Unit tests are implemented for the Enzyme class.

Signed-off-by: zoltuz <zoltuz@gmail.com>