Refactor Enzyme class (#212)
MultiEnzyme and Enzyme classes are merged because they had the same functionality. The signature of Enzyme has changed to substrates (plural) and products (plural), but it accepts a single string, Species or Component not just lists of those. Example notebooks are adjusted to the new signature.
Unit tests are implemented for the refactored Enzyme class.

Closes #148

Signed-off-by: zoltuz <zoltuz@gmail.com>