add Units to BioCRNpyler (#219)
This PR adds units compatibility for BioCRNpyler (addressing #130 ).  Summary of how it works:

units.py has a list of functions that create unit definitions (as required by SBML). These functions are factored in a way that new functions can be easily added with no changes required at any other place in the codebase.
So, to add new units, the user simply adds their units to their parameter file in a new column, adds a function corresponding to that unit id string in units.py (this is a simple 3 line function that can be created by copying one of the existing function for units).
Parameter class has a new attribute "unit" that goes with every Parameter object (we can discuss what's the best way to do this). This is the part that is incomplete and I need some help so that we can get this done quickly.
The tsv file Perfect Parameter File has a new column "units", we will use this to test the functionality to make sure units are correctly read and written to SBML.
So, the current to-do is a minor one: successfully read the Perfect Parameter file with units so that all parameters that have been read have a param.unit attribute.]

* Parameter units loaded from files and constructors with appropriate tests
* units working, 5. Parameter notebook has the example SBML using the perfect parameter csv
* add tests for units and unit list checks
* fix warning pytest check
* add a dictionary of supported units

Co-authored-by: William Poole <wp_ix@hotmail.com>
Co-authored-by: William Poole <wpoole@caltech.edu>
Co-authored-by: Eldad Afik <eldad.afik@gmail.com>
Co-authored-by: Zoltan Tuza <zoltuz@gmail.com>
Co-authored-by: dr3y <andrey.shur@gmail.com>