Simplifed Complex Function
In particular: Complex no longer works directly on OrderedPolymerSpecies to create PolymerConformations. Instead, OrderedPolymerSpecies must be placed inside a PolymerConformation in order to create new PolymerConformations. This simplifies the code by seperating the logic for PolymerSpecies and PolymerConformations. The downside is that Polymers are effectively static once inside a conformation - meaning bindign something to a polymer then putting it into a conformation is different from putting the polymer into a conformation then binding something to it. However, once everything is a polymer conformation order doesn't matter.